+++ SkiaSharpFiddle.Win/MainWindow.xaml
@@ -3,7 +3,7 @@
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:SkiaSharpFiddle"
+        xmlns:viewModel="clr-namespace:SkiaSharpFiddle.ViewModels;assembly=SkiaSharpFiddle"
         xmlns:skia="clr-namespace:SkiaSharp.Views.WPF;assembly=SkiaSharp.Views.WPF"
         xmlns:avalonEdit="http://icsharpcode.net/sharpdevelop/avalonedit"
         mc:Ignorable="d"
@@ -57,7 +57,7 @@
 
 
     <Window.DataContext>
-        <local:MainViewModel />
+        <viewModel:MainViewModel />
     </Window.DataContext>
 
     <DockPanel>
