+++ SkiaSharpFiddle/Compiler/AssemblyLoader.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 
-namespace SkiaSharpFiddle
+namespace SkiaSharpFiddle.Compiler
 {
     public class AssemblyLoader
     {
